created NucleusModule with animations, forms and pipes in it
