+++ src/platform/core/index.ts
@@ -11,6 +11,11 @@ import { MaterialModule } from '@angular/material';
  * COMPONENTS
  */
 
+// Nucleus
+
+import { CovalentNucleusModule } from './nucleus';
+export * from './nucleus';
+
 // Layouts
 import { TdLayoutComponent } from './layout/layout.component';
 import { TdLayoutNavComponent } from './layout/layout-nav/layout-nav.component';
@@ -116,67 +121,6 @@ export { TdConfirmDialogComponent } from './dialogs/confirm-dialog/confirm-dialo
 export { TdPromptDialogComponent } from './dialogs/prompt-dialog/prompt-dialog.component';
 
 /**
- * DIRECTIVES
- */
-
-import { TdToggleDirective } from './directives/toggle/toggle.directive';
-import { TdFadeDirective } from './directives/fade/fade.directive';
-import { TdAutoTrimDirective } from './directives/auto-trim/auto-trim.directive';
-
-export const TD_PLATFORM_DIRECTIVES: Type<any>[] = [
-  TdToggleDirective,
-  TdFadeDirective,
-  TdAutoTrimDirective,
-];
-
-export { TdToggleDirective } from './directives/toggle/toggle.directive';
-export { TdFadeDirective } from './directives/fade/fade.directive';
-export { TdAutoTrimDirective } from './directives/auto-trim/auto-trim.directive';
-
-/**
- * VALIDATORS
- */
-import { TdMinValidator } from './validators/min.validator';
-import { TdMaxValidator } from './validators/max.validator';
-import { TdNumberRequiredValidator } from './validators/number-required.validator';
-
-export const TD_VALIDATORS: Type<any>[] = [
-  TdMinValidator,
-  TdMaxValidator,
-  TdNumberRequiredValidator,
-];
-
-export { TdMinValidator } from './validators/min.validator';
-export { TdMaxValidator } from './validators/max.validator';
-export { TdNumberRequiredValidator } from './validators/number-required.validator';
-
-/**
- * PIPES
- */
-import { TdOrderByPipe } from './pipes/orderby/orderby.pipe';
-import { TdTimeAgoPipe } from './pipes/time-ago/time-ago.pipe';
-import { TdTimeDifferencePipe } from './pipes/time-difference/time-difference.pipe';
-import { TdBytesPipe } from './pipes/bytes/bytes.pipe';
-import { TdDigitsPipe } from './pipes/digits/digits.pipe';
-import { TdTruncatePipe } from './pipes/truncate/truncate.pipe';
-
-export const TD_PIPES: Type<any>[] = [
-  TdOrderByPipe,
-  TdTimeAgoPipe,
-  TdTimeDifferencePipe,
-  TdBytesPipe,
-  TdDigitsPipe,
-  TdTruncatePipe,
-];
-
-export { TdOrderByPipe } from './pipes/orderby/orderby.pipe';
-export { TdTimeAgoPipe } from './pipes/time-ago/time-ago.pipe';
-export { TdTimeDifferencePipe } from './pipes/time-difference/time-difference.pipe';
-export { TdBytesPipe } from './pipes/bytes/bytes.pipe';
-export { TdDigitsPipe } from './pipes/digits/digits.pipe';
-export { TdTruncatePipe } from './pipes/truncate/truncate.pipe';
-
-/**
  * MEDIA
  */
 import { TdMediaService } from './media/services/media.service';
@@ -185,11 +129,6 @@ import { TdMediaToggleDirective } from './media/directives/media-toggle.directiv
 export { TdMediaService } from './media/services/media.service';
 export { TdMediaToggleDirective } from './media/directives/media-toggle.directive';
 
-/**
- * ANIMATIONS
- */
-export { TdCollapseAnimation } from './animations/collapse.animation';
-
 @NgModule({
   imports: [
     HttpModule,
@@ -197,18 +136,16 @@ export { TdCollapseAnimation } from './animations/collapse.animation';
     FormsModule,
     CommonModule,
     MaterialModule.forRoot(),
+    CovalentNucleusModule.forRoot(),
   ],
   declarations: [
     TdMediaToggleDirective,
-    TD_PIPES,
     TD_LAYOUT_DIRECTIVES,
     TdLoadingDirective,
     TdLoadingComponent,
     TD_STEP_DIRECTIVES,
     TD_EXPANSION_DIRECTIVES,
     TD_DIALOG_DIRECTIVES,
-    TD_PLATFORM_DIRECTIVES,
-    TD_VALIDATORS,
   ],
   exports: [
     HttpModule,
@@ -216,17 +153,15 @@ export { TdCollapseAnimation } from './animations/collapse.animation';
     FormsModule,
     CommonModule,
     MaterialModule,
+    CovalentNucleusModule,
 
     TdMediaToggleDirective,
-    TD_PIPES,
     TD_LAYOUT_DIRECTIVES,
     TdLoadingDirective,
     TdLoadingComponent,
     TD_STEP_DIRECTIVES,
     TD_EXPANSION_DIRECTIVES,
     TD_DIALOG_DIRECTIVES,
-    TD_PLATFORM_DIRECTIVES,
-    TD_VALIDATORS,
   ],
   entryComponents: [ TD_DIALOG_ENTRY_COMPONENTS ],
 })
