+++ src/platform/core/nucleus/animations/collapse/collapse.animation.ts
