+++ src/platform/core/nucleus/animations/fade/fade.directive.ts
