+++ src/platform/core/nucleus/index.ts
@@ -1,0 +1,1 @@
+export * from './nucleus.module';
