+++ src/platform/core/nucleus/pipes/bytes/bytes.pipe.ts
