+++ src/platform/core/nucleus/pipes/truncate/truncate.pipe.spec.ts
