+++ src/platform/core/nucleus/pipes/truncate/truncate.pipe.ts
