+++ src/platform/core/steps/step-body/step-body.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 
 import { StepState } from '../step.component';
 
-import { TdCollapseAnimation } from '@covalent/core/animations/collapse.animation';
+import { TdCollapseAnimation } from '@covalent/core/nucleus/animations/collapse/collapse.animation';
 
 @Component({
   selector: 'td-step-body',
